move formatting functions to separate module and airFryerCost input field formatting and saving
